# apps/api/openapi.gen.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Char AI API",
    "description": "AI services API for speech-to-text transcription, LLM chat completions, and subscription management",
    "license": {
      "name": ""
    },
    "version": "1.0.0"
  },
  "paths": {
    "/calendar/google/list-calendars": {
      "post": {
        "tags": [
          "calendar"
        ],
        "operationId": "google_list_calendars",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GoogleListCalendarsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Google calendars fetched",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/google.ListCalendarsResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/calendar/google/list-events": {
      "post": {
        "tags": [
          "calendar"
        ],
        "operationId": "google_list_events",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GoogleListEventsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Google events fetched",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/google.ListEventsResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
    "/nango/connections": {
      "get": {
        "tags": [
          "nango"
        ],
        "operationId": "list_connections",
        "responses": {
          "200": {
            "description": "List of active connections",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListConnectionsResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "nango"
        ],
        "operationId": "delete_connection",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteConnectionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Connection disconnected",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteConnectionResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/nango/session": {
      "post": {
        "tags": [
          "nango"
        ],
        "operationId": "create_session",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSessionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Session created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "AccessRole": {
        "type": "string",
        "enum": [
          "freeBusyReader",
          "reader",
          "writer",
          "owner",
          "unknown"
        ]
      },
...
      "BirthdayProperties": {
        "type": "object",
        "properties": {
          "contact": {
            "type": [
              "string",
              "null"
            ]
          },
          "customTypeName": {
            "type": [
              "string",
              "null"
            ]
          },
          "type": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/BirthdayPropertyType"
              }
            ]
          }
        }
      },
...
      "CalendarListEntry": {
        "type": "object",
        "required": [
          "id"
        ],
        "properties": {
          "accessRole": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/AccessRole"
              }
            ]
          },
          "autoAcceptInvitations": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "backgroundColor": {
            "type": [
              "string",
              "null"
            ]
          },
          "colorId": {
            "type": [
              "string",
              "null"
            ]
          },
          "conferenceProperties": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ConferenceProperties"
              }
            ]
          },
          "dataOwner": {
            "type": [
              "string",
              "null"
            ]
          },
          "defaultReminders": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "$ref": "#/components/schemas/Reminder"
            }
          },
          "deleted": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "etag": {
            "type": [
              "string",
              "null"
            ]
          },
          "foregroundColor": {
            "type": [
              "string",
              "null"
            ]
          },
          "hidden": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "id": {
            "type": "string"
          },
          "kind": {
            "type": [
              "string",
              "null"
            ]
          },
          "location": {
            "type": [
              "string",
              "null"
            ]
          },
          "notificationSettings": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/NotificationSettings"
              }
            ]
          },
          "primary": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "selected": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "summary": {
            "type": [
              "string",
              "null"
            ]
          },
          "summaryOverride": {
            "type": [
              "string",
              "null"
            ]
          },
          "timeZone": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "ConferenceProperties": {
        "type": "object",
        "properties": {
          "allowedConferenceSolutionTypes": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "$ref": "#/components/schemas/ConferenceSolutionType"
            }
          }
        }
      },
...
      "ConnectionItem": {
        "type": "object",
        "required": [
          "integration_id",
          "connection_id"
        ],
        "properties": {
          "connection_id": {
            "type": "string"
          },
          "integration_id": {
            "type": "string"
          },
          "last_error_at": {
            "type": [
              "string",
              "null"
            ]
          },
          "last_error_description": {
            "type": [
              "string",
              "null"
            ]
          },
          "last_error_type": {
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "type": [
              "string",
              "null"
            ]
          },
          "updated_at": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "CreateSessionRequest": {
        "type": "object",
        "required": [
          "integration_id"
        ],
        "properties": {
          "connection_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "integration_id": {
            "type": "string"
          },
          "mode": {
            "$ref": "#/components/schemas/SessionMode"
          }
        }
      },
...
      "DeleteConnectionRequest": {
        "type": "object",
        "required": [
          "connection_id",
          "integration_id"
        ],
        "properties": {
          "connection_id": {
            "type": "string"
          },
          "integration_id": {
            "type": "string"
          }
        }
      },
      "DeleteConnectionResponse": {
        "type": "object",
        "required": [
          "status"
        ],
        "properties": {
          "status": {
            "type": "string"
          }
        }
      },
...
      "EventAttachment": {
        "type": "object",
        "properties": {
          "fileId": {
            "type": [
              "string",
              "null"
            ]
          },
          "fileUrl": {
            "type": [
              "string",
              "null"
            ]
          },
          "iconLink": {
            "type": [
              "string",
              "null"
            ]
          },
          "mimeType": {
            "type": [
              "string",
              "null"
            ]
          },
          "title": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "EventPerson": {
        "type": "object",
        "properties": {
          "displayName": {
            "type": [
              "string",
              "null"
            ]
          },
          "email": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": [
              "string",
              "null"
            ]
          },
          "self": {
            "type": [
              "boolean",
              "null"
            ]
          }
        }
      },
...
      "GoogleListCalendarsRequest": {
        "type": "object",
        "required": [
          "connection_id"
        ],
        "properties": {
          "connection_id": {
            "type": "string"
          }
        }
      },
      "GoogleListEventsRequest": {
        "type": "object",
        "required": [
          "connection_id",
          "calendar_id"
        ],
        "properties": {
          "calendar_id": {
            "type": "string"
          },
          "connection_id": {
            "type": "string"
          },
          "max_results": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "minimum": 0
          },
          "order_by": {
            "type": [
              "string",
              "null"
            ]
          },
          "page_token": {
            "type": [
              "string",
              "null"
            ]
          },
          "single_events": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "time_max": {
            "type": [
              "string",
              "null"
            ]
          },
          "time_min": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "ListConnectionsResponse": {
        "type": "object",
        "required": [
          "connections"
        ],
        "properties": {
          "connections": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ConnectionItem"
            }
          }
        }
      },
...
      "NotificationSettings": {
        "type": "object",
        "properties": {
          "notifications": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "$ref": "#/components/schemas/CalendarNotification"
            }
          }
        }
      },
...
      "Reminder": {
        "type": "object",
        "required": [
          "method",
          "minutes"
        ],
        "properties": {
          "method": {
            "$ref": "#/components/schemas/ReminderMethod"
          },
          "minutes": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "ReminderMethod": {
        "type": "string",
        "enum": [
          "email",
          "popup",
          "unknown"
        ]
      },
...
      "SessionMode": {
        "type": "string",
        "enum": [
          "auto",
          "connect",
          "reconnect"
        ]
      },
      "SessionResponse": {
        "type": "object",
        "required": [
          "token",
          "expires_at",
          "mode"
        ],
        "properties": {
          "connection_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "expires_at": {
            "type": "string"
          },
          "mode": {
            "type": "string"
          },
          "token": {
            "type": "string"
          }
        }
      },
...
      "google.Event": {
        "type": "object",
        "required": [
          "id"
        ],
        "properties": {
          "anyoneCanAddSelf": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "attachments": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "$ref": "#/components/schemas/EventAttachment"
            }
          },
          "attendees": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "$ref": "#/components/schemas/google.Attendee"
            }
          },
          "attendeesOmitted": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "birthdayProperties": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/BirthdayProperties"
              }
            ]
          },
          "colorId": {
            "type": [
              "string",
              "null"
            ]
          },
          "conferenceData": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ConferenceData"
              }
            ]
          },
          "created": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "creator": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/EventPerson"
              }
            ]
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "end": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/EventDateTime"
              }
            ]
          },
          "endTimeUnspecified": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "etag": {
            "type": [
              "string",
              "null"
            ]
          },
          "eventType": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/google.EventType"
              }
            ]
          },
          "extendedProperties": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ExtendedProperties"
              }
            ]
          },
          "focusTimeProperties": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/FocusTimeProperties"
              }
            ]
          },
          "gadget": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/Gadget"
              }
            ]
          },
          "guestsCanInviteOthers": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "guestsCanModify": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "guestsCanSeeOtherGuests": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "hangoutLink": {
            "type": [
              "string",
              "null"
            ]
          },
          "htmlLink": {
            "type": [
              "string",
              "null"
            ]
          },
          "iCalUID": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "string"
          },
          "kind": {
            "type": [
              "string",
              "null"
            ]
          },
          "location": {
            "type": [
              "string",
              "null"
            ]
          },
          "locked": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "organizer": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/EventPerson"
              }
            ]
          },
          "originalStartTime": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/EventDateTime"
              }
            ]
          },
          "outOfOfficeProperties": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/OutOfOfficeProperties"
              }
            ]
          },
          "privateCopy": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "recurrence": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            }
          },
          "recurringEventId": {
            "type": [
              "string",
              "null"
            ]
          },
          "reminders": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/Reminders"
              }
            ]
          },
          "sequence": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "source": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/EventSource"
              }
            ]
          },
          "start": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/EventDateTime"
              }
            ]
          },
          "status": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/EventStatus"
              }
            ]
          },
          "summary": {
            "type": [
              "string",
              "null"
            ]
          },
          "transparency": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/Transparency"
              }
            ]
          },
          "updated": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "visibility": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/Visibility"
              }
            ]
          },
          "workingLocationProperties": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/WorkingLocationProperties"
              }
            ]
          }
        }
      },
...
      "google.ListCalendarsResponse": {
        "type": "object",
        "properties": {
          "etag": {
            "type": [
              "string",
              "null"
            ]
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CalendarListEntry"
            }
          },
          "kind": {
            "type": [
              "string",
              "null"
            ]
          },
          "nextPageToken": {
            "type": [
              "string",
              "null"
            ]
          },
          "nextSyncToken": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "google.ListEventsResponse": {
        "type": "object",
        "properties": {
          "accessRole": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/AccessRole"
              }
            ]
          },
          "defaultReminders": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "$ref": "#/components/schemas/Reminder"
            }
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "etag": {
            "type": [
              "string",
              "null"
            ]
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/google.Event"
            }
          },
          "kind": {
            "type": [
              "string",
              "null"
            ]
          },
          "nextPageToken": {
            "type": [
              "string",
              "null"
            ]
          },
          "nextSyncToken": {
            "type": [
              "string",
              "null"
            ]
          },
          "summary": {
            "type": [
              "string",
              "null"
            ]
          },
          "timeZone": {
            "type": [
              "string",
              "null"
            ]
          },
          "updated": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Supabase JWT token"
      }
    }
  },
...
}